Move `PaymentElement` to its own file
